Move ValueFormat out of Value
Yes it makes sense to make it accompany the parser, because the parse acts on its specification.
However, such an abstraction should be composable then and clearly makes case for another library.
For now Value is used as an abstraction over the choice of value parser, otherwise we'd have to isolate that into another abstraction layer.